Implemented skills page
Added all of the skills, with their respective animation (Took me some time to figure them out) and their colors, those may change in the future if needed.

Added the component progressCircle for the skill overview.

Added new component for two paragraph animations

Added a new color to the vars, grey1

Added the new portfolio page, will be the next to do.