Added the portfolio section
Since i need some time to check what projects to put (or even if i could put one of my previous in there) i opted for putting a svg done by me (3 times for the different languages) and animating it! Could need some shadows to make it look a little better.

Fixed a bug where the icon arrow would translate for having the same class as the icon in language controller.

Added sass to the project for simplicity while adding the svg styles.

Added variables to most? colors (Have to check the whole project to see if there are more)

Added the new boilerplate for the contact page and added it to the router.