Added the final page, contact
Added some left margin classes on the app.vue file

Changed the way the progress circle behaves, it now shows all of the percentages when hovering one of the items inside, for better user experience. Also changed the way the circle loads, might change later after some tests.

Added the new contact page, with some svg icons newly made and a new neon button, with the working copy mechanism for the contact icons and the option to send mail directly. Might add an option to copy the mail instead of opening a mailto link.