Added the mobile navbar
Removed old dependency vue-animejs

Changed all of the animations to the normal animejs library instead of the removed one

Added a variation to the sidebar file, the mobile navbar. Added all of the animations and changed many styles to make it fit accordinlgy.

Also added some animations and a variable to the app file (windowSize)  with 3 specific values 0 (For 1/1 aspect ratio or less), 1 (for aspect ratio of 8/5) and 2 for the rest (Big wider screens)

Added some basic icons for the navbar

Changed the position of the languageController when the screen is of size 0, also the animations of it.