Created component mobileNavItem
Created new component that replaces the individual values set on the mabile navbar in the sidebar component (Should be renamed to navbar)

Updated language controller so the animations of resize play better according to the screen

Added a single new text in the en file

Fixed most of the page components having a problem with the fade animation with the router, where it would shrink too early in the animation.

Changed the way the portfolio sign behaves, not it displays each sign on top of each other and changes the visibility so all of them have the same swinging animation.