Created new component for the mobile skills
Created a new component progressBar so the skills can be displayed on mobile without distortion of any kind.

Removed overflow in router clases inside the app component. Removed the webkit-tap-highlitht so it now doesnt show a blue highlight when clicked a button (or link or div)

Changed some styles for the portfolio svg

Changed the skills page to add the new component for the mobile component. Also changed the way it renders the progressCircle and progressBar components, it now uses a vfor for simplicity and reusability. also added a transition to the divs surrounding the main blocks.

Changed some styles all around to make the mobile versions look better.