Replaced the old mobile navbar with new style
Need to see if the new mobile bar is better than the other one with some testing with users.

Changed some styles to make the app look a little bit better

Also tested on mobile to be confident that the styles worked correctly.

Deleted a variable in the sidebar that wasnt very usefull and changed the way it delays the router anim.