Added all the translations and some finishing touches
Added all the translations to all pages and components. Added a new favicon and a new title to the page.

Removed the notFound page since it doesnt serve any purpouse

Changed the color of the title to be a global change instead of a local change

Added a new script in bash to help push the changes inside and publish it to github pages, hopefully will work (The command is set up to be 'npm run deploy')